Call complete quest appropriately
